requirements.txt
flask
uvicorn
sqlalchemy
strawberry-graphql